Updates to allsubs, and connection
Moved the connection command to a function so if one things changes you don't have to change several places.
All return now return a json file so it looks pretty.
Allsubs in now now all generated so it does not need to be updated as more types of subs are added.